refactor: extract common logic on multiple and single plots
