+++ front/src/components/PlotJcamp/types.ts
@@ -1,0 +1,8 @@
+export interface PlotQuery {
+  xLabel: string;
+  yLabel: string;
+  xUnits: string;
+  yUnits: string;
+  scale: 'log' | 'linear';
+  logFilter?: 'remove' | 'abs';
+}
